moved voltage allocation to `mcm`
